index.html: show summary in a collapsible list
Yocto gathers the amount of CVEs per branch at the top of their metrics view.
However, the presentation of this information is not descriptive enough and it’s spread across several files.
This change adds collapsible, nested lists to show all cve information.

This commit shows current CVE count per release,
parse txt files with CVE lists to group them by project and display their total CVE count.

Inline this data on the page in details elements so there’s no need to navigate away.

Also, there will be another patch for the autobuilder-helper project where we update the data that is used.
It changes the format of the txt file so the metrics can handle it easily.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>