patch-status/index.html: use echarts to display all charts
Now we use echarts and add a separate controls section (zoom)
for this purpose and remove interactivity from the old charts themselves
With this change we only need to load one, the complete, set of data. The zoom option makes it easy to filter tha charts by date-ranges.

We also load releases.csv only once

Note that there will be another patch for the autobuilder-helper project where we update the data that is used.
Means that the jsonfiles with data for only the last year don't get created anymore.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>